requirements.txt
Faker>=18.0.0
PyYAML>=6.0
pandas>=1.5
duckdb>=1.0
